% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nb.r
\name{calculate_nb}
\alias{calculate_nb}
\title{Given the metadata and tag data, calculate Nb values.}
\usage{
calculate_nb(tags, generations = 1, metadata_column = "Nb", write = NULL)
}
\arguments{
\item{tags}{Input tags provided by count_tags.pl and hopefully count_otus.pl}

\item{generations}{Set the g parameter from the published paper.}
}
\description{
This is minor cleanup of the original implementation I received from
Dr. Lee's collaborators.  I do have the text of the original paper, but it is
difficult to interpret; so I fell back on basically retranscribing the
original code.
}
