% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{calculate_resiliency}
\alias{calculate_resiliency}
\title{Internal Resiliency Function.}
\usage{
calculate_resiliency(
  reference_vector,
  reads_df,
  cfu_vector,
  noiseless_table,
  sample_name,
  step_df,
  min_weight = 0.03,
  input_calibration = NULL,
  output_csv = "rtisan_values.csv",
  output_noiseless = "rtisan_noiseless.csv",
  innoculation_cfu = 1e+08,
  do_plots = FALSE,
  noise_correction = 0.005,
  verbose = FALSE
)
}
\arguments{
\item{reference_vector}{Set of reference reads/tag.}

\item{reads_df}{Table of the reads/tag for the samples of interest.}

\item{cfu_vector}{Set of cfu values from which to acquire the
target cfu for this particular sample.}

\item{noiseless_table}{Input table of noisless data, currently
unused.}

\item{sample_name}{Name of this samples, used to extract some data.}

\item{step_df}{Used to calculate the computed_bottleneck by
stats::NLSstClosestX and generate the Ns value.}

\item{min_weight}{Constant of the minimum weight!}

\item{input_calibration}{Input calibration file, currently unused.}

\item{output_csv}{Output csv file for writing new values.}

\item{innoculation_cfu}{Assumed cfu when not provided.}

\item{do_plots}{Print plots?}

\item{noise_correction}{Constant for correcting noise!}

\item{verbose}{Print text while running?}
}
\description{
This is run for every experimental or calibration sample in the
metadata.
}
