% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{convert_cutoffs}
\alias{convert_cutoffs}
\title{I do not really get this function, here is the author's text:}
\usage{
convert_cutoffs(p, output_vector, bind_df)
}
\arguments{
\item{p}{One less than q.}

\item{output_vector}{Definitely not the input_vector}

\item{bind_df}{the thing from which we will take the next item...}
}
\description{
This function takes the values in guessesunquesorted and
identifies their location in the plot of barcode frequencies vs
barcode, which is defined by cutoffpositions
}
