% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_samples.r
\name{exclude_samples}
\alias{exclude_samples}
\title{Plot the calibration curve of the tags observed vs. CFU}
\usage{
exclude_samples(
  tags,
  max_dilution = NULL,
  max_cfu = NULL,
  type = "calibration",
  sampleids = NULL,
  replicate = NULL
)
}
\arguments{
\item{tags}{Tag data.}

\item{x_column}{Metadata column for the x axis.}

\item{y_column}{Metadata column for the y axis.}

\item{color_column}{Metadata column for colors.}

\item{transform_x}{Perform a transformation on the x axis?}

\item{transform_y}{Perform a transformation on the y axis?}
}
\description{
Currently this is just a simple ggplot scatterplot.  I want to make it more
configurable and fun.
}
