% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_given_reference.r
\name{filter_given_reference}
\alias{filter_given_reference}
\title{Filter samples using the tags observed in the reference samples.}
\usage{
filter_given_reference(tag_list, reference_cutoff = 1, remove_nas = FALSE)
}
\arguments{
\item{reference_cutoff}{Minimum number of reads in the reference samples to
be considered 'real'}

\item{remove_nas}{Replace NA with 0?}

\item{index_list}{Result from count_tags}
}
\description{
Take the result from read_tags() and use it to find which tags are in the
non-reference samples.  One caveat about this process, it splits the data by
replicate because tags are not global.  Thus we will need to consider what is
the best way to handle the filtered data.  The simplest is to just merge the
pieces back together.
}
