% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_given_reference.r
\name{filter_given_reference_list}
\alias{filter_given_reference_list}
\title{This does the real work for filter_given_reference().}
\usage{
filter_given_reference_list(index_list, reference_cutoff = 1)
}
\arguments{
\item{index_list}{Result from count_tags}

\item{reference_cutoff}{Minimum number of reads in the reference samples to
be considered 'real'}
}
\description{
Create a list of matrices containing only the tags observed in each
replicates' reference sample.
}
