% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tags.r
\name{filter_tags}
\alias{filter_tags}
\title{Try out some heuristics for filtering the tag data.}
\usage{
filter_tags(tag_list, multiplier = 1)
}
\arguments{
\item{tag_list}{Result from one form of read_tags().  This contains the
metadata and the matrix of reads/tag.}

\item{multiplier}{Arbitrary multiplier to make the filter more stringent.}
}
\description{
Given our focus on RNAseq, it is not a stretch to guess that we would assume
some of those methods would be useful.  Here is a cpm implementation of some
simple filtering.
}
