% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_tags.r
\name{filter_topn_tags}
\alias{filter_topn_tags}
\title{Keep only the top-n most abundant tags}
\usage{
filter_topn_tags(tag_list, replicate_column = "replicate", topn = 600)
}
\arguments{
\item{tag_list}{Existing tag data}

\item{replicate_column}{Tags should be shared across replicates.}

\item{topn}{How many tags to keep.}
}
\description{
In Dr. Lee's experiment, they explicitly know the number of input tags.  So
we can explicitly keep only those n most observed tags.
}
