% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire1.R
\docType{data}
\name{fire1}
\alias{fire1}
\title{Forest Fire dataset}
\format{\code{fire1} --- a \code{SpatialPolygonsDataFrame} with 33 records of of the location of forest fire. \code{fire2} a..}
\source{
To be written
}
\description{
A dataset containing forest fire polygons representing the movement of the XX forest fire from T1 (fire1) to T2 (fire2). The extent of the fire on T1 was xx ha across yy polygons. By T2 the fire extent has increased/decreased to xx ha across yy polygons.
}
\details{
The \code{fire1} and \code{fire2} datases contains polygons that were derived from XX imagery... more info. the dataset.
}
\examples{
library(sp)
data(fire1)
data(fire2)
plot(fire1)
plot(fire2,border=2,add=TRUE)
}
\references{
Some References go here
}
\keyword{datasets}

