% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire2.R
\docType{data}
\name{fire2}
\alias{fire2}
\title{Forest Fire dataset}
\format{\code{fire2} --- a \code{SpatialPolygonsDataFrame} with 33 records of of the location of forest fire.}
\description{
see fire1
}
\keyword{datasets}

