% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_given_reference.r
\name{gather_tags_per_replicate}
\alias{gather_tags_per_replicate}
\title{Extract the tags observed in each replicate's reference sample.}
\usage{
gather_tags_per_replicate(meta, reads_per_tag, reference_cutoff = 1)
}
\arguments{
\item{meta}{Metadata matrix.}

\item{reads_per_tag}{The set of tags from the index list.}
}
\description{
This ought to be extended to multiple samples/reference.
}
