% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{get_interval_nb}
\alias{get_interval_nb}
\title{Grab the maximum value from a portion of the nb vector.}
\usage{
get_interval_nb(y, nb_vector = NULL)
}
\description{
Grab the maximum value from a portion of the nb vector.
}
