% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nb.r
\name{make_frequency_df}
\alias{make_frequency_df}
\title{The f(i,s) term in the Nb estimation equation requires frequency estimates.}
\usage{
make_frequency_df(mtrx)
}
\arguments{
\item{mtrx}{Matrix to transform.}
}
\description{
Thus, take the sum of each column and divide every value by it.
}
