% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpb.R
\docType{data}
\name{mpb}
\alias{mpb}
\title{MPB dataset}
\format{\code{mpb} --- a \code{SpatialPolygonsDataFrame} with 711 hotspot polygons that occured over eight years. The temporal indicator is the \code{TGROUP} column. Another variable \code{REGION} indicates whether the hotspot was in the northern or southern regions, which experienced mostly independent outbreaks.}
\source{
Data obtained from Trisalyn Nelson (ASU)
}
\description{
A dataset containing polygons representing the location of mountain pine beetle hotspot polygons in Morice Forest District, British Columbia, Canada.
}
\details{
These data were derived from helicopter-based GPS surveys during early years of large mountain pine beetle outbreak in Western Canada.
}
\examples{
library(sp)
data(mpb)
plot(mpb,border=2,add=TRUE)
}
\references{
Nelson TA, Boots B, Wulder MA, Carroll AL. Environmental characteristics of mountain pine beetle infestation hot spots. \emph{Journal of Ecosystems and Management}. 2007 Mar 14;8(1).
}
\keyword{datasets}

