% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nb.r
\name{nb_from_tags}
\alias{nb_from_tags}
\title{Given a tag data structure, calculate the Nb values and put them
into the metadata.}
\usage{
nb_from_tags(
  tags,
  generations = 1,
  metadata_column = "Nb",
  zero_filter = TRUE,
  write = NULL
)
}
\arguments{
\item{tags}{Data provided by read_csv_tags, read_qiime_otus, or
read_idx_tags.}

\item{generations}{Set the 'g' parameter from the original paper.}

\item{metadata_column}{Put the vector of Nb values into this new
column of the metadata.}

\item{zero_filter}{Filter out tags not observed in the reference
samples?}

\item{write}{Write the output to this file if not null.}
}
\value{
List containing a backup of the original metadata, new
 metadata, the column name with the new Nb values, the vector of Nb
 values, and the input tags.
}
\description{
The tag data structure contains the metadata along with the tags.
This function therefore passes the relevant pieces to
calculate_Nb() and sends the result into the metadata column of
choice.  If one specifies the write argument, it will write out a
copy of the new metadata.
}
