% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{nb_minimum_search}
\alias{nb_minimum_search}
\title{Search for minima...  Another function I don't understand.}
\usage{
nb_minimum_search(p, nb_vector)
}
\description{
Here is the text from the authors:
Outer layer of the function that scans for minima. The input to
this, p, is later defined. p is a specific set of numbers that
specifiy all the poisitions for which the minimum finder will
start. scanformin is applied over each of these positions.
}
\details{
Hmm yeah...
}
