% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.r
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot the calibration curve of the tags observed vs. CFU}
\usage{
plot_calibration(
  tags,
  x_column = "cfu",
  y_column = NULL,
  color_column = "replicate",
  transform_x = "log2",
  transform_y = "log2"
)
}
\arguments{
\item{tags}{Tag data.}

\item{x_column}{Metadata column for the x axis.}

\item{y_column}{Metadata column for the y axis.}

\item{color_column}{Metadata column for colors.}

\item{transform_x}{Perform a transformation on the x axis?}

\item{transform_y}{Perform a transformation on the y axis?}
}
\description{
Currently this is just a simple ggplot scatterplot.  I want to make it more
configurable and fun.
}
