% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_density.r
\name{plot_read_density}
\alias{plot_read_density}
\title{Make a boxplot of read density}
\usage{
plot_read_density(retlist)
}
\arguments{
\item{retlist}{from read_idx or read_qiime}
}
\description{
Make a boxplot of read density
}
