% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cfu.r
\name{predict_cfu}
\alias{predict_cfu}
\title{Try out some methods to go from tags/Nb back to CFU.}
\usage{
predict_cfu(meta, from = "Nb", to = "cfu", provided = NULL)
}
\arguments{
\item{meta}{Metadata.}

\item{from}{Factor from which to extrapolate.}

\item{to}{Factor to which to return.}

\item{provided}{Dataframe of incomplete data.}
}
\description{
This function has not been completed, mostly because I am not sure if it will
ever be used, but also because I am not sure of the appropriate model.  I
have never had an opportunity to use modelling to estimate data, so this is
mostly a place for me to play around with something that I should know but
don't.
}
