% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_reads.r
\name{preprocess_stamp}
\alias{preprocess_stamp}
\title{Preprocess raw reads using cutadapt and my little perl script.}
\usage{
preprocess_stamp(
  metadata,
  raw_column = "raw_fastq",
  trimmed_column = "trimmed_fastq",
  index_column = "index_table",
  output = NULL
)
}
\arguments{
\item{metadata}{Sample metadata.}

\item{column}{Metadata column containing the locations of the raw reads.}
}
\description{
Preprocess raw reads using cutadapt and my little perl script.
}
