% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_tags.r
\name{read_csv_tags}
\alias{read_csv_tags}
\title{Get tags from a csv file}
\usage{
read_csv_tags(
  input_csv,
  meta,
  id_column = "sampleid",
  cutoff = NULL,
  sample_column = "replicate"
)
}
\arguments{
\item{input_csv}{CSV file in the format used by the original authors.}

\item{meta}{Metadata in my format.}

\item{id_column}{Metadata column containing the samples by replicate..}

\item{cutoff}{I forget.}

\item{sample_column}{Metada column containing the parent sample IDs.}
}
\description{
This short circuits the processing functions.
}
