% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime_otus.r
\name{read_qiime_otus}
\alias{read_qiime_otus}
\title{Create data structures similar to read_idx using qiime otus.}
\usage{
read_qiime_otus(
  metadata,
  otu_column = "qiime_otus",
  xref_sequence = FALSE,
  trimmed_column = "trimmed_reads",
  output = NULL
)
}
\arguments{
\item{metadata}{Sample metadata.}

\item{output}{Write the matrix to this file, if provided.}

\item{column}{Metadata column containing the qiime output.}
}
\description{
There are some problems with this implementation still, primarily because the
version of qiime does not helpfully give the tag sequences.  I have found
some ways around this, but since I don't really like that implementation I
haven't finished it yet.
}
