% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idx.r
\name{read_tags}
\alias{read_tags}
\title{Interpret the files produced by my count_idx.pl script.}
\usage{
read_tags(
  metadata,
  id_column = "sampleid",
  index_column = "index_table",
  output = NULL,
  cutoff = 3
)
}
\arguments{
\item{metadata}{Sample metadata.}

\item{id_column}{Column in the metadata containing the sample names.}

\item{index_column}{Column in the metadata containing the tag tables.}

\item{output}{Write out the matrix to this file (if provided).}

\item{cutoff}{Initial reads/tag filter.}
}
\description{
This uses the index_table column in the metadata, reads this files associated
with it, and creates some data structures with the results.  These include a
modified version of the metadata, containing some summary information, a
table of the reads/tag observed, and a long table for ggplot.
}
