% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{rtisan_nb_from_tags}
\alias{rtisan_nb_from_tags}
\title{A partial copying of the getNrNb() function from rtisan.}
\usage{
rtisan_nb_from_tags(
  tags,
  generations = 1,
  metadata_column = "pre_rtisan_Nb",
  zero_filter = FALSE,
  write_metadata = NULL,
  innoculation_cfu = 1e+08,
  chosen_seed = 1,
  min_weight = 0.03,
  noise_correction = 0.005,
  input_calibration = NULL,
  output_csv = "rtisan_values.csv",
  output_noiseless = "rtisan_noiseless.csv",
  do_plots = FALSE,
  verbose = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{tags}{The data structure generated by read_idx or
read_csv_tags or read_qiime_otus.  E.g. the list of metadata, tags
by sample, and some administrativia.}

\item{generations}{Used by calculate_nb}

\item{metadata_column}{Used by nb_from_tags to add the pre-rtisan
Nb values to the metadata.}

\item{zero_filter}{Used by calculate_nb to leave off tags not
observed in the reference samples. I think this is slightly broken
though.}

\item{innoculation_cfu}{Used by rtisan, defines a constant for the
innoculation cfu in case it was not provided in the metadata.}

\item{chosen_seed}{Set the seed for extraDistr so that can get the
same result two times in a row.}

\item{min_weight}{The minimum weight constant, not sure what it does yet.}

\item{noise_correction}{Noise correction constant, not sure what it
does yet.}

\item{input_calibration}{Input csv file containing calibration
constants.  I do not have one, so I cannot say how it works yet.}

\item{output_csv}{Write the new Nb values to this file: FIXME Move
this to the metadata.}

\item{do_plots}{Print the diagnostic plots?  FIXME: ggplot-ify
these and potentially get them into the metadata.}

\item{verbose}{Print some diagnostics while running?}

\item{write_metadat}{When not null, write out the changes to the
metadata to the provided file.}

\item{output_noisless}{Write the noiseless values to this file,
FIXME: also put this into the metadata.}
}
\description{
This function takes in the tags data structure I have been using
and passes the data to a combination of the Nb calculator I have
along with the resiliency filtered Nb values.
}
\details{
I do not yet fully understand the various functions/operations
performed for the resiliency filtering, as a result, these
functions are mostly just transcriptions of the originals with some
variables renamed and changes to make it easier to read.
}
