% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{rtisan_nr_nb}
\alias{rtisan_nr_nb}
\title{This is the top-level function from the authors.  I have it
wrapped.}
\usage{
rtisan_nr_nb(
  read_csv,
  cfu_vector,
  innoculation_cfu = 1e+08,
  chosen_seed = 1,
  reference_samples = c(1, 2, 3),
  min_weight = 0.03,
  noise_correction = 0.005,
  input_calibration = NULL,
  do_plots = FALSE,
  verbose = FALSE,
  output_csv = "rtisan_values.csv",
  output_noiseless = "rtisan_noiseless.csv",
  cpus = NULL,
  parallel = TRUE
)
}
\arguments{
\item{read_csv}{CSV file containing the reads/tag, FIXME: use the
tag data structure here to simplify things.}

\item{cfu_vector}{CFU values for each sample of interest.}

\item{innoculation_cfu}{Assumed cfu if they are not provided.}

\item{chosen_seed}{Set the seed you maniacs!}

\item{reference_samples}{Vector of reference samples, provided by
the metadata via the parent function.}

\item{min_weight}{Constant of minimum weight...}

\item{noise_correction}{Constant of noise correction...}

\item{input_calibration}{CSV of input calibrations.}

\item{do_plots}{Plot the results?}

\item{verbose}{Be verbose?}

\item{output_csv}{Print the new Nb values here.}

\item{output_noiseless}{Print the noiseless data here...}
}
\description{
It sets up the data and iterates over every experimental sample in
order to run the calculate resliency function upon it.
}
