% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{sorted_nb}
\alias{sorted_nb}
\title{Calculate Nb values on a set of sorted read/tag vectors.}
\usage{
sorted_nb(bindsorted, input_vector, output_vector, generations = 1)
}
\description{
This should trivially be able to be deleted in favor of calculate_nb().
}
