% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stampr-package.r
\docType{package}
\name{stampr-package}
\alias{stampr-package}
\title{stampr: Spatial Temporal Analysis of Moving Polygons}
\description{
The Package \code{stampr} provides tools for performing spatial temporal analysis of moving polygons.
These tools allow the calculation of directional relationships, distance relations, and other basic
functionality, such as global change metrics. More details about each of these functions
can be found in its help documentation.
}
\details{
\code{stampr}'s functions utilize the \code{sf} objects (as of version 0.3.0) from the 
package \code{sf}. Polygon relationships are still understudied in the field of 
geographic information science, but hopefully \code{stampr} can provide users with a
platform for new developments and applied research looking at interesting geographical phenomena.
}
\references{
Robertson, C., Nelson, T., Boots, B., and Wulder, M. (2007) STAMP: Spatial-temporal analysis of moving polygons
 \emph{Journal of Geographical Systems}, 9:207-227.
Long, J., Robertson, C., Nelson, T. (2018) stampr: Spatial-Temporal Analysis of Moving Polygons in R \emph{Journal of Statistical Software}. Code Snippets, 84(1), 1–19.
}
\author{
Jed Long and Colin Robertson
}
