% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtisan.r
\name{vector_proportion}
\alias{vector_proportion}
\title{Calculate the proprotion of each value in the vector with a twist!}
\usage{
vector_proportion(t, output_vector = NULL)
}
\arguments{
\item{t}{Number of elements on the bottom of the sorted vector from
which to calculate the sum which is used as the denominator of the
frequencies.}

\item{output_vector}{The input of reads/tag from which to calculate
this peculiar variant of frequencies.}
}
\description{
This is _almost_ a vector version of make_frequency_df, except it
takes only the last t values of the vector for the sum.  Why?  I am
not certain.
}
